+++ tests/Beanstalk/JobQueueTest.php
@@ -28,7 +28,7 @@ class JobQueueTest extends TestCase
      */
     protected $jobQueue;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
         $this->beanstalk = $this->createMock(ConnectionInterface::class);
@@ -36,7 +36,7 @@ class JobQueueTest extends TestCase
         $this->jobQueue = new JobQueue($this->beanstalk, $this->scheduler);
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         $this->jobQueue = null;
         $this->scheduler = null;
@@ -44,12 +44,12 @@ class JobQueueTest extends TestCase
         parent::tearDown();
     }
 
-    public function testIsInstanceOfJobQueueInterface()
+    public function testIsInstanceOfJobQueueInterface(): void
     {
         static::assertInstanceOf(JobQueueInterface::class, $this->jobQueue);
     }
 
-    public function testPutForImmediateJobCallsBeanstalk()
+    public function testPutForImmediateJobCallsBeanstalk(): void
     {
         $jobId = 123;
         $this->scheduler->expects(static::once())
@@ -63,13 +63,19 @@ class JobQueueTest extends TestCase
             ->willReturn($jobId);
 
         $job = $this->createMock(JobInterface::class);
+        $job->method('getDelay')
+            ->willReturn(rand(1, 100));
+
         static::assertEquals($jobId, $this->jobQueue->put($job));
     }
 
-    public function testPutForProlongedJobCallsScheduler()
+    public function testPutForProlongedJobCallsScheduler(): void
     {
         $jobId = 123;
         $job = $this->createMock(JobInterface::class);
+
+        $job->method('getDelay')
+            ->willReturn(rand(1, 100));
 
         $this->scheduler->expects(static::once())
             ->method('shouldBeScheduled')
@@ -82,7 +88,7 @@ class JobQueueTest extends TestCase
         static::assertEquals($jobId, $this->jobQueue->put($job));
     }
 
-    public function testRetrieveSuccessfully()
+    public function testRetrieveSuccessfully(): void
     {
         $jobId = 123;
         $body = [
@@ -98,7 +104,7 @@ class JobQueueTest extends TestCase
         static::assertEquals($jobId, $this->jobQueue->retrieve('testQueue')->getId());
     }
 
-    public function testRetrieveWhenNoJobsAvailable()
+    public function testRetrieveWhenNoJobsAvailable(): void
     {
         $this->beanstalk->expects(static::once())
             ->method('reserve')
@@ -109,7 +115,7 @@ class JobQueueTest extends TestCase
     /**
      * @expectedException \Phlib\JobQueue\Exception\InvalidArgumentException
      */
-    public function testRetrieveWithBadlyFormedBeanstalkData()
+    public function testRetrieveWithBadlyFormedBeanstalkData(): void
     {
         $this->beanstalk->expects(static::once())
             ->method('reserve')
@@ -120,7 +126,7 @@ class JobQueueTest extends TestCase
     /**
      * @expectedException \Phlib\JobQueue\Exception\JobRuntimeException
      */
-    public function testRetrieveWithBadlyFormedJobBody()
+    public function testRetrieveWithBadlyFormedJobBody(): void
     {
         $this->beanstalk->expects(static::once())
             ->method('reserve')
@@ -135,7 +141,7 @@ class JobQueueTest extends TestCase
      * @param mixed $jobData
      * @dataProvider jobDataMaintainsExpectedTypeDataProvider
      */
-    public function testJobDataMaintainsExpectedType($jobData)
+    public function testJobDataMaintainsExpectedType($jobData): void
     {
         $package = JobFactory::serializeBody(new Job('TestQueue', $jobData));
         $this->beanstalk->expects(static::once())
@@ -148,7 +154,7 @@ class JobQueueTest extends TestCase
         static::assertEquals($jobData, $job->getBody());
     }
 
-    public function jobDataMaintainsExpectedTypeDataProvider()
+    public function jobDataMaintainsExpectedTypeDataProvider(): array
     {
         return [
             [[
@@ -165,45 +171,57 @@ class JobQueueTest extends TestCase
         ];
     }
 
-    public function testMarkAsCompleteDeletesBeanstalkJob()
+    public function testMarkAsCompleteDeletesBeanstalkJob(): void
     {
         $jobId = 123;
         $job = $this->createMock(JobInterface::class);
         $job->expects(static::once())
             ->method('getId')
             ->willReturn($jobId);
+
         $this->beanstalk->expects(static::once())
             ->method('delete')
-            ->with($jobId);
+            ->with($jobId)
+            ->willReturn($this->beanstalk);
+
         $this->jobQueue->markAsComplete($job);
     }
 
-    public function testMarkAsIncompleteReleasesBeanstalkJobWhenDelayIsMoreImmediate()
+    public function testMarkAsIncompleteReleasesBeanstalkJobWhenDelayIsMoreImmediate(): void
     {
         $jobId = 123;
         $job = $this->createMock(JobInterface::class);
         $job->expects(static::once())
             ->method('getId')
             ->willReturn($jobId);
+        $job->method('getDelay')
+            ->willReturn(rand(1, 100));
+
         $this->scheduler->expects(static::once())
             ->method('shouldBeScheduled')
             ->willReturn(false);
+
         $this->beanstalk->expects(static::once())
             ->method('useTube')
             ->willReturnSelf();
         $this->beanstalk->expects(static::once())
             ->method('release')
-            ->with($jobId);
+            ->with($jobId)
+            ->willReturn($this->beanstalk);
+
         $this->jobQueue->markAsIncomplete($job);
     }
 
-    public function testMarkAsIncompleteReleasesBeanstalkJobWhenDelayIsMoreProlonged()
+    public function testMarkAsIncompleteReleasesBeanstalkJobWhenDelayIsMoreProlonged(): void
     {
         $jobId = 123;
         $job = $this->createMock(JobInterface::class);
         $job->expects(static::atLeastOnce())
             ->method('getId')
             ->willReturn($jobId);
+        $job->method('getDelay')
+            ->willReturn(rand(1, 100));
+
         $this->scheduler->expects(static::once())
             ->method('shouldBeScheduled')
             ->willReturn(true);
@@ -213,7 +231,7 @@ class JobQueueTest extends TestCase
         $this->jobQueue->markAsIncomplete($job);
     }
 
-    public function testMarkAsErrorBuriesBeanstalkJob()
+    public function testMarkAsErrorBuriesBeanstalkJob(): void
     {
         $jobId = 123;
         $job = $this->createMock(JobInterface::class);
@@ -222,7 +240,9 @@ class JobQueueTest extends TestCase
             ->willReturn($jobId);
         $this->beanstalk->expects(static::once())
             ->method('bury')
-            ->with($jobId);
+            ->with($jobId)
+            ->willReturn($this->beanstalk);
+
         $this->jobQueue->markAsError($job);
     }
 }